split aster
