+++ dev/services/wms/ows_refactored/aster/abstract_aster_cfg.py
@@ -1,0 +1,519 @@
+abstract_aster_false_colour = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+False colour RGB composite
+
+- Red: B3
+- Green: B2
+- Blue: B1
+(red = green vegetation)
+
+Use this image to help understand non-geological differences within and between ASTER scenes caused by green vegetation (red), fire scars, thin and thick cloud and cloud shadows.
+
+Use band 2 only for a gray-scale background to the content, composition and index colour products.
+
+For 'False Colour Mosaic' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74348
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_regolith_ratios = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+3 band RGB composite
+
+- Red: B3/B2
+- Green: B3/B7
+- Blue: B4/B7
+(white = green vegetation)
+
+Use this image to help interpret:
+
+(1) the amount of green vegetation cover (appears as white);
+
+(2) basic spectral separation (colour) between different regolith and geological units and regions/provinces; and
+
+(3) evidence for unmasked cloud (appears as green).
+
+For 'Regolith Ratios' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74349
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_aloh_group_composition = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: B5/B7
+
+- Blue is well ordered kaolinite, Al-rich muscovite/illite, paragonite, pyrophyllite
+- Red is Al-poor (Si-rich) muscovite (phengite)
+
+Useful for mapping:
+
+(1) exposed saprolite/saprock is often white mica or Al-smectite (warmer colours) whereas transported materials are often kaolin-rich (cooler colours);
+
+(2) clays developed over carbonates, especially Al-smectite (montmorillonite, beidellite) will produce middle to warmers colours;
+
+(3) stratigraphic mapping based on different clay-types; and
+
+(4) lithology-overprinting hydrothermal alteration, e.g. Si-rich and K-rich phengitic mica (warmer colours).
+
+Combine with Ferrous iron in MgOH and FeOH content products to look for evidence of overlapping/juxtaposed potassic metasomatism in ferromagnesian parents rocks (e.g. Archaean greenstone associated Au mineralisation) +/- associated distal propyllitic alteration (e.g. chlorite, amphibole)
+
+For 'AlOH Group Composition' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74356
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_aloh_group_content = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: (B5+B7)/B6
+
+- Blue is low abundance
+- Red is high abundance
+
+(potentially includes: phengite, muscovite, paragonite, lepidolite, illite, brammalite, montmorillonite, beidellite, kaolinite, dickite)
+
+Useful for mapping:
+
+(1) exposed saprolite/saprock;
+
+(2) clay-rich stratigraphic horizons;
+
+(3) lithology-overprinting hydrothermal phyllic (e.g. white mica) alteration; and
+
+(4) clay-rich diluents in ore systems (e.g. clay in iron ore).
+
+Also combine with AlOH composition to help map:
+
+(1) exposed in situ parent material persisting through “cover” which can be expressed as:
+
+(a) more abundant AlOH content +
+
+(b) long-wavelength (warmer colour) AlOH composition (e.g. muscovite/phengite)
+
+For 'AlOH Group Content' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74355
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_feoh_group_content = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: (B6+B8)/B7
+
+- Blue is low content,
+- Red is high content
+
+(potentially includes: chlorite, epidote, jarosite, nontronite, gibbsite, gypsum, opal-chalcedony
+
+Useful for mapping:
+
+(1) jarosite (acid conditions) – in combination with ferric oxide content (high);
+
+(2) gypsum/gibbsite – in combination with ferric oxide content (low);
+
+(3) magnesite - in combination with ferric oxide content (low) and MgOH content (moderate-high);
+
+(4) chlorite (e.g. propyllitic alteration) – in combination with Ferrous in MgOH (high); and
+
+(5) epidote (calc-silicate alteration) – in combination with Ferrous in MgOH (low).
+
+For 'FeOH Group Content' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74358
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_ferric_oxide_composition = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: B2/B1
+
+- Blue-cyan is goethite rich,
+- Green is hematite-goethite,
+- Red-yellow is hematite-rich
+
+Useful For:
+
+(1) Mapping transported materials (including palaeochannels) characterised by hematite (relative to geothite). Combine with AlOH composition to find co-located areas of hematite and poorly ordered kaolin to map transported materials; and
+
+(2) hematite-rish areas in drier conditions (eg above the water table) whereas goethite-rich in wetter conditions (eg at/below the water or areas recently exposed). May also be climate driven.
+
+For 'Ferric Oxide Composition' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74352
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_ferric_oxide_content = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: B4/B3
+
+- Blue is low abundance,
+- Red is high abundance
+
+Useful for:
+
+(1) Exposed iron ore (hematite-goethite). Use in combination with the “Opaques index” to help separate/map dark:
+
+   (a) surface lags (e.g. maghemite gravels) which can be misidentified in visible and false colour imagery; and
+
+   (b) magnetite in BIF and/or bedded iron ore; and
+
+(2) Acid conditions: combine with FeOH Group content to help map jarosite which will have high values in both products.
+
+Mapping hematite versus goethite mapping is NOT easily achieved as ASTER’s spectral bands were not designed to capture diagnostic iron oxide spectral behaviour.
+
+However, some information on visible colour relating in part to differences in hematite and/or goethite content can be obtained using a ratio of B2/B1 especially when this is masked using a B4/B3 to locate those pixels with sufficient iro oxide content.
+
+For 'Ferric Oxide Content' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74351
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_ferrous_iron_content_in_mgoh = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: B5/B4
+
+- Blue is low ferrous iron content in carbonate and MgOH minerals like talc and tremolite.
+- Red is high ferrous iron content in carbonate and MgOH minerals like chlorite and actinolite.
+
+Useful for mapping:
+
+(1) un-oxidised “parent rocks” – i.e. mapping exposed parent rock materials (warm colours) in transported cover;
+
+(2) talc/tremolite (Mg-rich – cool colours) versus actinolite (Fe-rich – warm colours);
+
+(3) ferrous-bearing carbonates (warm colours) potentially associated with metasomatic “alteration”;
+
+(4) calcite/dolomite which are ferrous iron-poor (cool colours); and
+
+(5) epidote, which is ferrous iron poor (cool colours) – in combination with FeOH content product (high).
+
+For 'Ferrous Iron Content in MgOH/Carbonate' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74361
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_ferrous_iron_index = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: B5/B4
+
+- Blue is low abundance,
+- Red is high abundance
+
+This product can help map exposed “fresh” (un-oxidised) rocks (warm colours) especially mafic and ultramafic lithologies rich in ferrous silicates (e.g. actinolite, chlorite) and/or ferrous carbonates (e.g. ferroan dolomite, ankerite, siderite).
+
+Applying an MgOH Group content mask to this product helps to isolate ferrous bearing non-OH bearing minerals like pyroxenes (e.g. jadeite) from OH-bearing or carbonate-bearing ferrous minerals like actinolite or ankerite, respectively.
+
+Also combine with the FeOH Group content product to find evidence for ferrous-bearing chlorite (e.g. chamosite).
+
+For 'Ferrous Iron Index' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74353
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_green_vegetation = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: B3/B2
+
+- Blue is low content,
+- Red is high content
+
+Use this image to help interpret the amount of “obscuring/complicating” green vegetation cover.
+
+For 'Green Vegetation Content' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74350
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_gypsum_index = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band Ratio: (B10+B12)/B11
+
+- Blue is low gypsum content,
+- Red is high gypsum content
+
+Useful for mapping:
+
+(1) evaporative environments (e.g. salt lakes) and associated arid aeolian systems (e.g. dunes);
+
+(2) acid waters (e.g. from oxidising sulphides) invading carbonate rich materials including around mine environments; and
+
+(3) hydrothermal (e.g. volcanic) systems.
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_kaolin_group_index = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band Ratio: B6/B5
+
+- Blue is low content,
+- Red is high content
+
+(potentially includes: pyrophyllite, alunite, well-ordered kaolinite)
+
+Useful for mapping:
+
+(1) different clay-type stratigraphic horizons;
+
+(2) lithology-overprinting hydrothermal alteration, e.g. high sulphidation, “advanced argillic” alteration comprising pyrophyllite, alunite, kaolinite/dickite; and
+
+(3) well-ordered kaolinite (warmer colours) versus poorly-ordered kaolinite (cooler colours) which can be used for mapping in situ versus transported materials, respectively.
+
+For 'Kaolin Group Index' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74357
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_mgoh_group_composition = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: B7/B8
+
+- Blue-cyan is magnesite-dolomite, amphibole, chlorite
+- Red is calcite, epidote, amphibole
+
+Useful for mapping:
+
+(1) exposed parent material persisting through "cover";
+
+(2) "dolomitization" alteration in carbonates - combine with Ferrous iron in MgOH product to help separate dolomite versus ankerite;
+
+(3) lithology-cutting hydrothermal (e.g. propyllitic) alteration - combine with FeOH content product and ferrous iron in Mg-OH to isolate chlorite from actinolite versus talc versus epidote; and
+
+(4) layering within mafic/ultramafic intrusives
+
+
+For 'MgOH Group Composition' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74360
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_mgoh_group_content = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: (B6+B9)/(B7+B8)
+
+- Blue is low content,
+- Red is high content
+
+(potentially includes: calcite, dolomite, magnesite, chlorite, epidote, amphibole, talc, serpentine)
+
+Useful for mapping:
+
+(1) “hydrated” ferromagnesian rocks rich in OH-bearing tri-octahedral silicates like actinolite, serpentine, chlorite and talc;
+
+(2) carbonate-rich rocks, including shelf (palaeo-reef) and valley carbonates(calcretes, dolocretes and magnecretes); and
+
+(3) lithology-overprinting hydrothermal alteration, e.g. “propyllitic alteration” comprising chlorite, amphibole and carbonate.
+
+The nature (composition) of the silicate or carbonate mineral can be further assessed using the MgOH composition product.
+
+For 'MgOH Group Content' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74359
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_opaque_index = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: B1/B4
+
+- Blue is low abundance,
+- Red is high abundance
+
+(potentially includes carbon black (e.g. ash), magnetite, Mn oxides, and sulphides in unoxidised environments)
+
+Useful for mapping:
+
+(1) magnetite-bearing rocks (e.g. BIF);
+
+(2) maghemite gravels;
+
+(3) manganese oxides;
+
+(4) graphitic shales.
+
+Note: (1) and (4) above can be evidence for “reduced” rocks when interpreting REDOX gradients.
+
+Combine with AlOH group Content (high values) and Composition (high values) products, to find evidence for any invading “oxidised” hydrothermal fluids which may have interacted with reduced rocks evident in the Opaques index product.
+
+For 'Opaque Index' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74354
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_silica_index = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: B13/B10
+
+- Blue is low silica content,
+- Red is high silica content
+
+(potentially includes Si-rich minerals, such as quartz, feldspars, Al-clays)
+
+Geoscience Applications:
+
+Broadly equates to the silica content though the intensity (depth) of this reststrahlen feature is also affected by particle size &lt;250 micron.
+
+Useful product for mapping:
+
+(1) colluvial/alluvial materials;
+
+(2) silica-rich (quartz) sediments (e.g. quartzites);
+
+(3) silification and silcretes; and
+
+(4) quartz veins.
+
+Use in combination with quartz index, which is often correlated with the Silica index.
+
+For 'TIR Silica index' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74362
+
+For service status information, see https://status.dea.ga.gov.au
+"""
+
+abstract_aster_quartz_index = """
+The National ASTER Map of Australia is the parent datafile of a dataset that comprises a set of 14+ geoscience products made up of mosaiced ASTER (Advanced Spaceborne Thermal Emission and Reflection Radiometer) scenes across Australia.
+
+ASTER calibration, processing and standardisation approaches have been produced as part of a large multi-agency project to facilitate uptake of these techniques and make them easily integrated with other datasets in a GIS.
+
+Collaborative research, undertaken by Geoscience Australia, the Commonwealth Scientific Research Organisation (CSIRO) and state and industry partners, on the world-class Mt Isa mineral province in Queensland was completed in 2008 as a test-case for these new methods.
+
+For parent datafile information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74347
+
+
+Band ratio: B11/(B10+B12)
+
+- Blue is low quartz content,
+- Red is high quartz content
+
+Geoscience Applications:
+
+Use in combination with Silica index to more accurately map “crystalline” quartz rather than poorly ordered silica (e.g. opal), feldspars and compacted clays.
+
+For 'TIR Quartz Index' dataset information, see the dataset record: http://pid.geoscience.gov.au/dataset/ga/74363
+
+For service status information, see https://status.dea.ga.gov.au
+"""
