+++ dev/services/wms/ows_refactored/aster/band_aster_cfg.py
@@ -1,0 +1,9 @@
+bands_aster = {
+    "Band_1": [],
+    "Band_2": [],
+    "Band_3": [],
+}
+
+bands_aster_single_band = {
+    "Band_1": [],
+}
